+++ src/qiqf/utility/array/index.ts
@@ -26,3 +26,51 @@ export function arrayChunk(args: InterfaceUtil.IArrayAny, size: number): Interfa
 export function compact(args: InterfaceUtil.IArrayAny): InterfaceUtil.IArrayAny {
     return args.filter(Boolean)
 }
+
+/**
+ * 计算数组中值出现的次数
+ */
+const INITIAL_NUMBER = 0
+export function countOccurrences(args: InterfaceUtil.IArrayAny, value: number) {
+    return args.reduce((count, arg) => {
+        return arg === value ? count + 1 : count
+    }, INITIAL_NUMBER)
+}
+
+/**
+ * 扁平化数组
+ * @param args
+ */
+export function deepFlatten(args: InterfaceUtil.IArrayAny): InterfaceUtil.IArrayAny[] {
+    return [].concat(...args.map(arg => Array.isArray(arg) ? deepFlatten(arg) : arg))
+}
+
+/**
+ * 返回两个数组之间的差异部分组成的数组
+ * @param arg1
+ * @param arg2
+ */
+export function difference(arg1: InterfaceUtil.IArrayAny, arg2: InterfaceUtil.IArrayAny): InterfaceUtil.IArrayAny[] {
+    return arg2.filter(v2 => !arg1.includes(v2))
+}
+
+/**
+ * 返回两个数组中相同的部分组成的数组
+ * @param arg1
+ * @param arg2
+ */
+export function sameArray(arg1: InterfaceUtil.IArrayAny, arg2: InterfaceUtil.IArrayAny): InterfaceUtil.IArrayAny[] {
+    return arg2.filter(v2 => arg1.includes(v2))
+}
+
+/**
+ * 数组去重
+ * @param args
+ */
+export function distinctValuesOfArray(args: InterfaceUtil.IArrayAny): InterfaceUtil.IArrayAny[] {
+    return [...new Set(args)];
+}
+
+
+
+